Extract per-class skill casting into GameLogic::Combat
Move the remaining skill-execution functions (SkillWarrior, SkillElf, the
UseSkill* family, the Ragefighter set and CastWarriorSkill, ~1,200 lines) out
of ZzzInterface.cpp into GameLogic/Combat/SkillCast.{h,cpp} under namespace
GameLogic::Combat. Its callers were already in that namespace (ClassAttack,
SkillExecution) so they needed only the new include; ZzzInterface's own callers
are now qualified. This clears the skill subsystem out of ZzzInterface.cpp.
No behaviour change.
