+++ src/source/Engine/Object/ZzzInterface.h
@@ -124,17 +124,8 @@ bool CheckTile(CHARACTER* c, OBJECT* o, float Range);
 void LetHeroStop(CHARACTER* c = NULL, BOOL bSetMovementFalse = FALSE);
 void SendCharacterMove(unsigned short Key, float Angle, unsigned char PathNum, unsigned char* PathX, unsigned char* PathY, unsigned char TargetX, unsigned char TargetY);
 void Attack(CHARACTER* c);
-bool SkillElf(CHARACTER* c, ITEM* p);
 void SendRequestMagic(int Type, int Key);
 void SendRequestMagicContinue(int Type, int x, int y, int Angle, BYTE Dest, BYTE Tpos, WORD TKey, BYTE* pSkillSerial);
-bool SkillWarrior(CHARACTER* c, ITEM* p);
-void UseSkillWarrior(CHARACTER* c, OBJECT* o);
-void UseSkillWizard(CHARACTER* c, OBJECT* o);
-void UseSkillElf(CHARACTER* c, OBJECT* o);
-void UseSkillSummon(CHARACTER* pCha, OBJECT* pObj);
-void UseSkillRagefighter(CHARACTER* pCha, OBJECT* pObj);
-void AttackRagefighter(CHARACTER* pCha, int nSkill, float fDistance);
-bool UseSkillRagePosition(CHARACTER* pCha);
 bool CheckTarget(CHARACTER* c);
 void MouseRButtonReset();
 BYTE GetDestValue(int xPos, int yPos, int xDst, int yDst);
